pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "salesshortcut"
version = "0.1.0"
description = "Multi-agent AI-powered SDR system built with Dedalus ADK"
requires-python = ">=3.11"
dependencies = [
    "dedalus_labs",
    "fastapi>=0.110",
    "uvicorn[standard]>=0.29",
    "websockets>=12.0",
    "python-dotenv>=1.0",
    "pydantic>=2.6",
    "httpx>=0.27",
    "tenacity>=8.2",
    "jinja2>=3.1",
    "google-cloud-bigquery>=3.17",
    "google-api-python-client>=2.120",
    "google-auth>=2.28",
    "google-auth-oauthlib>=1.2",
    "google-cloud-pubsub>=2.19",
    "elevenlabs>=1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "ruff>=0.3",
]
